requirements.txt
peewee==3.16.0
psycopg2==2.9.6
python-dotenv==1.0.0
telebot==0.0.5